Remove build-essential cookbook
Since chef 14, the cookbook is merged in chef, but does not allow an
easy way to enable a specific repo when installing needed packages. This
was making installation of `kernel-devel` fail as Criteo dedicated
kernel was in a separate and disabled by default repo.

Simple solution is to remove build-essential cookbook from zookeeper
cookbook, and manually install needed dependencies.