♾️ Fix periodic rewrite of static conf by chef
This is because when dynamicReconfig is enabled, Zookeeper rewrites its
static conf with new dynamicConfigFile file, but it also reorders the
other keys. So we now keep order of keys, and append the new ones at the
end (they will be reorder when reconfig will be called once).

JIRA: MESOS-4478