Fix cluster initialization on dynamic_config
During new cluster initialization we've observed a parsing bug after the
second chef-client run. It was due to dynamic config missing the "client
port" part:
> server.11=mesos-master11-par.central.criteo.preprod:2888:3888:participant

instead of what would be expected:
> server.11=mesos-master11-par.central.criteo.preprod:2888:3888:participant;0.0.0.0:2181

This made parsing fail. When launching the reconfiguration once
(manually) with exactly the same value, dynamic configuration was
written in the correct (full) format by zookeeper.

It means for some reason zookeeper does not write the dynamic
configuration always in the same format and we need to support this.
We think this bug was not seen before because we were always in the
following cases:
- either migrating (manually) from static to dynamic config
- or bootstrapping the cluster, facing issues, fixing them manually

This patch also adds nice logs about dynamic config change because it
would have helped us understanding faster the issue.

Change-Id: I0893b2f9c5579ab39488f5768fd605b5103e9fcb